+++ crates/holochain/src/core/workflow/app_validation_workflow.rs
@@ -130,6 +130,7 @@ use holochain_state::prelude::*;
 
 use parking_lot::Mutex;
 use rusqlite::Transaction;
+use std::cmp::max;
 use std::collections::HashSet;
 use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::Ordering;
@@ -189,17 +190,17 @@ pub async fn app_validation_workflow(
         trigger_integration.trigger(&"app_validation_workflow");
     }
 
+    let validations_dependencies = validation_dependencies.lock();
     Ok(
         // If not all ops have been validated
         // and fetching missing hashes has not timed out,
         // trigger app validation workflow again.
         if outcome_summary.validated < outcome_summary.ops_to_validate
-            && !validation_dependencies
-                .lock()
-                .fetch_missing_hashes_timed_out()
+            && !validations_dependencies.fetch_missing_hashes_timed_out()
         {
-            // Trigger app validation workflow again in 10 seconds.
-            WorkComplete::Incomplete(Some(Duration::from_secs(10)))
+            // Trigger app validation workflow again in 20-2000 milliseconds.
+            let interval = max(101 - validations_dependencies.missing_hashes.len(), 1) * 20;
+            WorkComplete::Incomplete(Some(Duration::from_millis(interval as u64)))
         } else {
             WorkComplete::Complete
         },